Rename Rep.Mngr.SphericalExpansion to ...SphericalGTOExpansion
The existing SphericalExpansion class has so much code specific to the
GTO radial basis that the DVR implementation is best done in a new
class; this also implies that the existing class had to be renamed to
reflect that it's not the _only_ spherical expansion class.

Be sure to merge this change into the relevant SOAP branches ASAP:
If you try to merge into master after this change has gone through, the
shit will hit the fan.